src/requirements.txt
Flask
mysql-connector-python
spacy
rapidfuzz
scikit-learn
tensorflow
nltk
requests
jsonify
